+++ sdk/eventhubs/azure_messaging_eventhubs_checkpointstore_blob/src/checkpoint_store.rs
@@ -238,15 +238,11 @@ impl CheckpointStore for BlobCheckpointStore {
                     .map(|pos| &name[pos + 1..])
                     .unwrap_or_default()
                     .to_string();
-                if let Some(additional_properties) = blob
-                    .metadata
-                    .as_ref()
-                    .and_then(|m| m.additional_properties.as_ref())
-                {
-                    if let Some(sequence_number) = additional_properties.get(SEQUENCE_NUMBER) {
+                if let Some(values) = blob.metadata.as_ref().and_then(|m| m.values.as_ref()) {
+                    if let Some(sequence_number) = values.get(SEQUENCE_NUMBER) {
                         checkpoint.sequence_number = Some(sequence_number.parse()?);
                     }
-                    if let Some(offset) = additional_properties.get(OFFSET) {
+                    if let Some(offset) = values.get(OFFSET) {
                         checkpoint.offset = Some(offset.clone());
                     }
                 }
@@ -303,8 +299,8 @@ impl CheckpointStore for BlobCheckpointStore {
                 ownership.owner_id = blob
                     .metadata
                     .as_ref()
-                    .and_then(|m| m.additional_properties.as_ref())
-                    .and_then(|ap| ap.get(OWNER_ID).cloned());
+                    .and_then(|m| m.values.as_ref())
+                    .and_then(|v| v.get(OWNER_ID).cloned());
             }
             if let Some(properties) = &blob.properties {
                 ownership.etag = properties.etag.clone();